add sqlite3 to project, add some code to work it
This is supposed to be a cache for user information, so that <struct
user_info> can be shared from a GCache backed by a sqlite database, permitting
useful reception of updates with the "trim_user" option set.

The naming and file management strategy and whatnot are silly, but can be
changed later so fuck it for now.