refactor the tweet_view backing model
Now there's just a single column with the update in it, and the column
renderer functions grab the same object and pick different things out. It
should rather be that `struct user_info' was converted into a PtUserInfo
GObject, which would be stuck in the model's first column as appropriate from
a retweet/oc point of view. The reference counting will aid in effective
caching.